Make "default" more similar and `backend_opts(*, type=, priortize=, disable=)`
In a sense the main thing is to further the `type=`, thought.

This isn't perfect, but I am starting to think the problem of trying
to fit a 2-D problem into 1-D is best solved by keeping it 2-D :).
`type=` means we can make handle the type decision independently
from any backend priority decision.