Create an initial PoC implementation for `spatch` (#3)
This is an initial and relatively full featured dispatching implementation.
However, it is also very early, with many API decisions not going through detailed
review.

I.e. right now it is still clearly exploratory, it may not change much, or it might change a lot!

I'll keep some of the individual commits details around below :):

WIP: Start working based on Erik's and my old protypes and new ideas

* default backend has no types (for now) but ignore any unknown ones

* Introduce default primary types for the fallback

We need this in order to simplify rejecting numpy arrays for a
cupy array.
(This way, we don't have to reject it explicitly, but explicitly
opt into supporting it.  Which doesn't mean we won't need reject
matching if certain types are found.)

* Add PoC/rudimentary backend prioritization (type hierarchy + I am better)

* Add priority context manager

* make it a backend context manager in the "library"

* allow relevant args=None to dispatch on everything

* Enable passing of `DispatchInfo` also the callable

This amy be very relevant if we have explicitly `.invoke()` style
API, since that case the passed types can be different from the
actual types.

It also passes "prioritized" for information (although it should be
mainly important for `should_run`).

* Use topological sort at initialization and cache backend matching phase

* Automatically generate functions info in entry_points

* Make "default" more similar and `backend_opts(*, type=, priortize=, disable=)`

In a sense the main thing is to further the `type=`, thought.

This isn't perfect, but I am starting to think the problem of trying
to fit a 2-D problem into 1-D is best solved by keeping it 2-D :).
`type=` means we can make handle the type decision independently
from any backend priority decision.

* Add basic call tracing (maybe more to show that it can be done).

* add some docs, woohoo!

* Fix backend priority comparisons and report cycles/inconsistencies

* Create manual context manager for backend opts (better functionality and docs)

* fix "prioritized" and some other smaller changes/additions

* enable environment var config (not sure if best choice)

* Replace disable with BLOCK env that never loads the entry point

We need a way to not load one ever (probably also one to make only
an allowed list!).  OTOH, disable seems not really so important
so skip it for now

* rename to context and smaller cleanups

and maybe some re-org that not everyone likes :)

* Add design choice docs (that will probably get outdated, but ok...)

And some tiny doc build fixes

* add subclass and abstract class support (not tested much though)

* Add require_opt_in mechanism, BackendOpts repr and some cleanups

Also rename to prioritize_above and prioritize_below

* rename based on Erik's suggestion

* Fix toposort and some other things and add very basic tests

(This needs re-organizing to move some of the helping stuff
into a helper file probably, but some tests there are.)

* Make should-run True/False strict to ease adding "maybe"

While it is easy enough to add, I don't really want to add
a should run right now.  So instead, just make things strict
so that it would be easy to add (even without an API break).

* Add CI (stripped down) CI and gitignore changes from cookie cutter

Also fixes some things and enables coverage runs and test the entry
point generation to achieve a decent coverage

* Apply suggestions from code review

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>
Co-authored-by: Erik Welch <erik.n.welch@gmail.com>
Co-authored-by: Aditi Juneja <91629733+Schefflera-Arboricola@users.noreply.github.com>