ENH,MAINT: Rename "relevant" to "dispatch" args and expose in context (#13)
This is useful for generic wrappers.  I still wan to hold back to do
all the work for them, but the information is rather readily available,
so I think we can pass it on.

NumPy uses "relevant args" naming for the arguments we dispatch on,
I decided to rename it to "dispatch" instead.
(Also changed to pass a tuple of types, because I think that is more
future proof.  We expect few types, a set isn't useful e.g. if done in C.)