ENH,BUG: Allow dict and callable "dispatch_args" specification (#14)
This is a bit unwieldy.  I had an old start where I used the signature
(and some of the docs still said so), but that wasn't actually the case.

This now properly documents the fact that a `dict` can be passed
(which comes pretty directly from NetworkX).
Unlike networkx, a single parameter is not assumed to be parameter 0
but instead, `inspect.signature` is used to figure this out.

`dispatch_args` can now also be a callable, which is more clearly
defined.