Make `@dispatchable` return a function (#19)
Not returning a function can confuse some users (because they have strict checks for `FunctionType`).  NumPy got away with this, but it could be one stumbling block.

NOTE(seberg): In the future, with some things moved to C, this could be a more significant overhead and we may want to consider optionally disabling it as an optimization (or more carefully checking how necessary it is and changing the default).