MAINT: Remove any imports from `__init__` (hide `__version__`)
The helpers were just not needed and should be private.

`__version__` is a bit awkard and probably not quite right. Hiding
it seems good to me, I also changed it to look-up via `__name__`.

The thing is, I would like to vendor it, and then looking up `spatch`
is wrong and can fail. Looking up `__name__` may not actually be any
better, but for now I'll assume it is better than what we have.

Not sure if this matters either way, if it does we might have to find
a better solution...